requirements.txt
streamlit>=1.28.0
letsfg>=0.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.0
pytest>=7.0.0